auto delete wormholes on copy paste sigs, do not delete kspace to kspace wormholes, they never have sigs..

+++ modules/scanning/classes/controller/Scanner.class.php
@@ -221,7 +221,10 @@ namespace scanning\controller
 							{
 								$connection = $sig->getWormhole()->getConnectionTo($wormhole->solarSystemID);
 								if ($connection != null)
-									$connection->delete();
+                                {
+                                    if (!$connection->isKspaceToKspace())
+                                        $connection->delete();
+                                }
 							}
 						}
 					}
